VORTEX-5636 : Spreadsheet function refactor. (#24)
* VORTEX-5636 : Initial.

* Move enum to separate file.

* Refactoring.

* Getters.
